// frontend/src/app/shared/proxy/models/TrackingTopPageResult.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type TrackingTopPageResult = {
  trackingPageId?: string;
  slug?: string | null;
  title?: string | null;
  totalVisits?: number;
  uniqueVisitors?: number;
};